SiteUser 엔티티 추가, 다른 엔티티의 주키타입도 Long 으로 변경

+++ src/main/java/com/ll/exam/sbb/question/QuestionService.java
@@ -27,7 +27,7 @@ public class QuestionService {
     }
 
 
-    public Question getQuestion(int id) {
+    public Question getQuestion(long id) {
         return questionRepository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException("no %d question not found,".formatted(id)));
     }
